backend/pyproject.toml
# PopupChatKit Backend
# Python 后端项目配置文件
# 包管理工具: uv

[project]
name = "popupchatkit-backend"
version = "1.0.0"
description = "PopupChatKit 后端服务 - 基于 FastAPI 和 pydantic-ai 的 AI 对话系统"
authors = [
    {name = "Colin", email = "your-email@example.com"}
]
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["ai", "chatbot", "fastapi", "pydantic-ai", "llm"]

# 项目依赖
dependencies = [
    # Web 框架
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6", # 表单数据解析
    # AI 相关
    "pydantic-ai>=0.0.7",
    "openai>=1.12.0", # OpenAI SDK (pydantic-ai 依赖)
    # 数据处理
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    # 工具库
    "httpx>=0.26.0", # HTTP 客户端
    "aiofiles>=23.2.1", # 异步文件操作
    "python-jose[cryptography]>=3.3.0", # JWT 加密
    "passlib[bcrypt]>=1.7.4", # 密码加密
    # 日志和监控 (可选)
    "logfire>=0.23.0", # Pydantic Logfire (可选)
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
# 开发依赖
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",            # 测试 HTTP 请求
    "black>=23.12.0",           # 代码格式化
    "ruff>=0.1.9",              # 代码检查
    "mypy>=1.8.0",              # 类型检查
]





